requirements.txt
langchain_classic==1.0.0
langchain_community==0.4.1
langchain_core==1.2.2
langchain_google_genai==4.1.1
langchain_huggingface==1.2.0
langchain_text_splitters==1.1.0
protobuf==6.33.2
pypdf==6.4.2
python-dotenv==1.2.1
Requests==2.32.5
streamlit==1.48.1
supabase==2.27.0
